+++ app/src/main/java/com/google/codelab/hotpepperapiapp/ext/FragmentExt.kt
@@ -1,7 +1,8 @@
-package com.google.codelab.hotpepperapiapp
+package com.google.codelab.hotpepperapiapp.ext
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
+import com.google.codelab.hotpepperapiapp.R
 
 object FragmentExt {
     fun showFragment(fragmentManager: FragmentManager, fragment: Fragment) {
