+++ app/src/main/java/com/google/codelab/hotpepperapiapp/view/PagerStoreAdapter.kt
@@ -1,4 +1,4 @@
-package com.google.codelab.hotpepperapiapp
+package com.google.codelab.hotpepperapiapp.view
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,6 +6,8 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.google.codelab.hotpepperapiapp.R
+import com.google.codelab.hotpepperapiapp.model.Store
 
 class PagerStoreAdapter(private val store: List<Store>, private val listener: ListListener) :
     RecyclerView.Adapter<PagerViewHolder>() {
