+++ app/src/main/res/layout/fragment_favorite_store.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FavoriteStoreFragment">
+    tools:context=".view.FavoriteStoreFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_view_favorite"
