+++ app/src/main/res/layout/fragment_maps.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/frameLayout"
-    tools:context=".MapsFragment">
+    tools:context=".view.MapsFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
@@ -19,7 +19,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            tools:context=".MapsFragment" />
+            tools:context=".view.MapsFragment" />
 
         <androidx.viewpager2.widget.ViewPager2
             android:id="@+id/store_pager"
