+++ app/src/main/res/layout/fragment_store_web_view.xml
@@ -13,7 +13,7 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".StoreWebViewFragment">
+        tools:context=".view.StoreWebViewFragment">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
